+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 <h1 align="center">Changelog :</h1>
 
+### Version 1.0.1-Beta
+
+- Debug compotition to GPU
+- Change swapinterval : 90
+- Add render engine enabler
+- Add FPS Throttling [Here](https://developer.android.com/games/gamemode/gamemode-interventions#fps_throttling)
+
 ### Version 1.0-Beta
 
 - Initial Build v1.0-Beta